+++ src/Schemable.ts
@@ -8,7 +8,7 @@
  *
  * @since 2.2.0
  */
-import { Kind, URIS, HKT } from 'fp-ts/lib/HKT'
+import { HKT, Kind, Kind2, URIS, URIS2 } from 'fp-ts/lib/HKT'
 
 /**
  * @since 2.2.0
@@ -56,6 +56,32 @@ export interface Schemable1<S extends URIS> {
 }
 
 /**
+ * @since 2.2.8
+ */
+export interface Schemable2C<S extends URIS2, E> {
+  readonly URI: S
+  readonly literal: <A extends readonly [Literal, ...Array<Literal>]>(...values: A) => Kind2<S, E, A[number]>
+  readonly string: Kind2<S, E, string>
+  readonly number: Kind2<S, E, number>
+  readonly boolean: Kind2<S, E, boolean>
+  readonly nullable: <A>(or: Kind2<S, E, A>) => Kind2<S, E, null | A>
+  readonly type: <A>(properties: { [K in keyof A]: Kind2<S, E, A[K]> }) => Kind2<S, E, { [K in keyof A]: A[K] }>
+  readonly partial: <A>(
+    properties: { [K in keyof A]: Kind2<S, E, A[K]> }
+  ) => Kind2<S, E, Partial<{ [K in keyof A]: A[K] }>>
+  readonly record: <A>(codomain: Kind2<S, E, A>) => Kind2<S, E, Record<string, A>>
+  readonly array: <A>(items: Kind2<S, E, A>) => Kind2<S, E, Array<A>>
+  readonly tuple: <A extends ReadonlyArray<unknown>>(
+    ...components: { [K in keyof A]: Kind2<S, E, A[K]> }
+  ) => Kind2<S, E, A>
+  readonly intersect: <B>(right: Kind2<S, E, B>) => <A>(left: Kind2<S, E, A>) => Kind2<S, E, A & B>
+  readonly sum: <T extends string>(
+    tag: T
+  ) => <A>(members: { [K in keyof A]: Kind2<S, E, A[K]> }) => Kind2<S, E, A[keyof A]>
+  readonly lazy: <A>(id: string, f: () => Kind2<S, E, A>) => Kind2<S, E, A>
+}
+
+/**
  * @since 2.2.3
  */
 export interface WithUnknownContainers<S> {
@@ -69,6 +95,14 @@ export interface WithUnknownContainers<S> {
 export interface WithUnknownContainers1<S extends URIS> {
   readonly UnknownArray: Kind<S, Array<unknown>>
   readonly UnknownRecord: Kind<S, Record<string, unknown>>
+}
+
+/**
+ * @since 2.2.8
+ */
+export interface WithUnknownContainers2C<S extends URIS2, E> {
+  readonly UnknownArray: Kind2<S, E, Array<unknown>>
+  readonly UnknownRecord: Kind2<S, E, Record<string, unknown>>
 }
 
 /**
@@ -90,6 +124,15 @@ export interface WithUnion1<S extends URIS> {
 }
 
 /**
+ * @since 2.2.8
+ */
+export interface WithUnion2C<S extends URIS2, E> {
+  readonly union: <A extends readonly [unknown, ...Array<unknown>]>(
+    ...members: { [K in keyof A]: Kind2<S, E, A[K]> }
+  ) => Kind2<S, E, A[number]>
+}
+
+/**
  * @since 2.2.3
  */
 export interface WithRefine<S> {
@@ -101,6 +144,16 @@ export interface WithRefine<S> {
  */
 export interface WithRefine1<S extends URIS> {
   readonly refine: <A, B extends A>(refinement: (a: A) => a is B, id: string) => (from: Kind<S, A>) => Kind<S, B>
+}
+
+/**
+ * @since 2.2.8
+ */
+export interface WithRefine2C<S extends URIS2, E> {
+  readonly refine: <A, B extends A>(
+    refinement: (a: A) => a is B,
+    id: string
+  ) => (from: Kind2<S, E, A>) => Kind2<S, E, B>
 }
 
 /**